Annoteur-automatique/requirements.txt
pillow
numpy
scikit-learn
webcolors
pika